requirements.txt
streamlit
faiss-cpu
torch
numpy
sentence-transformers
transformers
pypdf
python-docx
fpdf